src/main/java/org/brightwheel/dto/DeviceReading.java
package org.brightwheel.dto;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import lombok.*;
import lombok.extern.jackson.Jacksonized;

import java.text.DateFormat;
import java.text.SimpleDateFormat;
import java.util.Date;
import java.util.TimeZone;

@Jacksonized
@Builder
@JsonIgnoreProperties(ignoreUnknown = true)
@ToString
@EqualsAndHashCode
public class DeviceReading {

    private Date timestamp;
    @Getter
    @EqualsAndHashCode.Exclude private int count;

    public String getTimestamp() {
        // TODO don't hardcode timezone
        TimeZone timeZone = TimeZone.getTimeZone("UTC");
        DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm'Z'");
        dateFormat.setTimeZone(timeZone);
        return dateFormat.format(timestamp);
    }
}
